Added file game_class
